test(pipeline): keep the lock-name helper test-only and fix a cast
lock_name_for has no production caller any more (claim_processing_in builds
the name from its explicit directory), so it is honestly cfg(test) rather than
kept alive for the compiler. Drops the redundant u64 cast clippy flagged.

+++ src-tauri/src/pipeline.rs
@@ -88,6 +88,10 @@ fn locks_dir() -> PathBuf {
     crate::paths::data_dir().join("locks")
 }
 
+/// اسم القفل للملف — **مستعمل في الاختبارات وحدها**: مسار الإنتاج يمرّ بـ
+/// `claim_processing_in` الذي يبني الاسم من مجلده الصريح، ولا ينادي هذه أبداً.
+/// والوسم هنا **صادق** لا تسكيت: لا مستدعي في المنتج.
+#[cfg(test)]
 fn lock_name_for(input: &Path) -> PathBuf {
     lock_name_in(&locks_dir(), input)
 }
@@ -831,7 +835,7 @@ mod tests {
             "الاسترجاع فوري، وقياسه {reclaim:?}"
         );
         assert!(
-            (reclaim.as_secs() as u64) < STALE_LOCK_SECS,
+            reclaim.as_secs() < STALE_LOCK_SECS,
             "الاسترجاع لم ينتظر الطابع الزمني (١٢ ساعة)"
         );
         eprintln!(
